pyproject.toml
[project]
name = "offbyone"
version = "0.1.0"
description = "Prez Discord Bot"
authors = [
    {name = "Prez",email = "154857421+P-R-E-Z@users.noreply.github.com"}
]
package-mode = false
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "discord-py (>=2.3.2,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
